Documents/work/서울경제신문/sedaily_chatbot0721/csc-genai-lab-main 6/packages/cdk_infra/tools/data_preprocessing/requirements.txt
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0